Refactor FEACN codes update job
